+++ web/app/connectors/_merlins-connector/account/commands.js
@@ -4,10 +4,10 @@
 
 import {makeRequest, makeFormEncodedRequest} from 'progressive-web-sdk/dist/utils/fetch-utils'
 import {jqueryResponse} from 'progressive-web-sdk/dist/jquery-response'
+import {getCurrentProductId} from 'progressive-web-sdk/dist/store/products/selectors'
 import {extractPathFromURL} from 'progressive-web-sdk/dist/utils/utils'
 import {SubmissionError} from 'redux-form'
 import {browserHistory} from 'progressive-web-sdk/dist/routing'
-
 import {getCookieValue, splitFullName} from '../../../utils/utils'
 import {getFormKey, getUenc} from '../selectors'
 import {fetchPageData} from '../app/commands'
@@ -17,7 +17,9 @@ import {
     setSigninLoaded,
     setRegisterLoaded,
     receiveAccountInfoData,
-    receiveAccountOrderListData
+    receiveAccountOrderListData,
+    receiveUpdatedWishlistItem,
+    removeWishlistItem
 } from 'progressive-web-sdk/dist/integration-manager/account/results'
 
 import {receiveCheckoutLocations} from 'progressive-web-sdk/dist/integration-manager/checkout/results'
@@ -35,7 +37,16 @@ import {
     parseAccountLocations
 } from './parsers'
 import {jqueryAjaxWrapper} from '../utils'
-import {CART_URL, LOGIN_POST_URL, CREATE_ACCOUNT_POST_URL, getDeleteAddressURL} from '../config'
+import {
+    CART_URL,
+    LOGIN_POST_URL,
+    CREATE_ACCOUNT_POST_URL,
+    getDeleteAddressURL,
+    UPDATE_WISHLIST_URL,
+    WISHLIST_URL,
+    getWishlistQuantityUrl
+} from '../config'
+
 import {setLoggedIn} from 'progressive-web-sdk/dist/integration-manager/results'
 
 export const initLoginPage = (url) => (dispatch) => {
@@ -81,7 +92,7 @@ export const initWishlistPage = (url) => (dispatch) => {
         .then(([$, $response]) => dispatch(receiveWishlistResponse($, $response)))
 }
 
-export const addToCartFromWishlist = ({itemId, productId, quantity}) => (dispatch, getState) => {
+export const addToCartFromWishlist = (productId, {itemId, quantity}) => (dispatch, getState) => {
     const currentState = getState()
     const formKey = getFormKey(currentState)
     const uenc = getUenc(productId)(currentState)
@@ -117,8 +128,14 @@ export const addToCartFromWishlist = ({itemId, productId, quantity}) => (dispatc
         })
 }
 
-export const removeItemFromWishlist = () => (dispatch) => Promise.resolve()
-
+export const removeItemFromWishlist = (itemId) => (dispatch, getState) => {
+    const requestBody = {
+        item: parseInt(itemId),
+        form_key: getFormKey(getState())
+    }
+    return makeFormEncodedRequest('/wishlist/index/remove/', requestBody, {method: 'POST'})
+        .then(() => dispatch(removeWishlistItem(itemId)))
+}
 
 const MAGENTO_MESSAGE_COOKIE = 'mage-messages'
 const clearMessageCookie = () => {
@@ -330,5 +347,35 @@ export const reorderPreviousOrder = (orderNumber) => (dispatch, getState) => { /
     const formKey = getFormKey(getState())
     const orderId = orderNumber.replace(/^0+/, '')
     return makeFormEncodedRequest(`/sales/order/reorder/order_id/${orderId}/`, {form_key: formKey}, {method: 'POST'})
-        .then(() => CART_URL)
+        .then(() => CART_URL) 
+}
+
+export const updateWishlistItem = (itemId, wishlistId, quantity) => (dispatch, getState) => {
+    const currentState = getState()
+    const productId = getCurrentProductId(currentState)
+    const payload = {
+        product: productId,
+        qty: quantity,
+        id: itemId,
+        form_key: getFormKey(currentState),
+        // This won't always be defined, but add to wishlist will still work
+        // if it's missing
+        uenc: getUenc(productId)(currentState)
+    }
+
+    return makeFormEncodedRequest(UPDATE_WISHLIST_URL, payload, {method: 'POST'})
+        .then(() => dispatch(receiveUpdatedWishlistItem({itemId, quantity})))
+        .then(() => WISHLIST_URL)
+}
+
+export const updateWishlistItemQuantity = (quantity, itemId, wishlistId) => (dispatch, getState) => {
+    const formKey = getFormKey(getState())
+    const requestBody = {
+        form_key: formKey,
+        do: '',
+    }
+    requestBody[`qty[${itemId}]`] = quantity
+    requestBody[`description[${itemId}]`] = ''
+    return makeFormEncodedRequest(getWishlistQuantityUrl(wishlistId), requestBody, {method: 'POST'})
+        .then(() => dispatch(receiveUpdatedWishlistItem({itemId, quantity})))
 }